requirements.txt
torch
torchvision
opencv-python
numpy
matplotlib
Pillow
seaborn
pandas
scikit-learn
networkx


